pkg/querier/querysharding/series.go
package querysharding

import (
	"github.com/cortexproject/cortex/pkg/querier/queryrange"
	"github.com/cortexproject/cortex/pkg/querier/series"
	"github.com/pkg/errors"
	"github.com/prometheus/common/model"
	"github.com/prometheus/prometheus/pkg/labels"
	"github.com/prometheus/prometheus/promql"
	"github.com/prometheus/prometheus/storage"
)

// needed to map back from api response to the underlying series data
func ResponseToSeries(resp queryrange.Response) (storage.SeriesSet, error) {
	switch resp.ResultType {
	case promql.ValueTypeVector, promql.ValueTypeMatrix:
		return newSeriesSet(resp.Result), nil
	}

	return nil, errors.Errorf(
		"Invalid promql.Value type: [%s]. Only %s and %s supported",
		resp.ResultType,
		promql.ValueTypeVector,
		promql.ValueTypeMatrix,
	)
}

func newSeriesSet(results []queryrange.SampleStream) storage.SeriesSet {

	set := make([]storage.Series, 0, len(results))

	for _, stream := range results {
		samples := make([]model.SamplePair, 0, len(stream.Samples))
		for _, sample := range stream.Samples {
			samples = append(samples, model.SamplePair{
				Timestamp: model.Time(sample.TimestampMs),
				Value:     model.SampleValue(sample.Value),
			})
		}

		ls := make([]labels.Label, 0, len(stream.Labels))
		for _, l := range stream.Labels {
			ls = append(ls, labels.Label(l))
		}
		set = append(set, series.NewConcreteSeries(ls, samples))
	}
	return series.NewConcreteSeriesSet(set)
}
